few changes

+++ app/src/main/java/com/revosleap/wpdroid/ui/recyclerview/models/post/PostResponse.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class PostResponse(
     @SerializedName("author")
-    var author: Int?,
+    var author: Long?,
     @SerializedName("categories")
     var categories: List<Long?>?,
     @SerializedName("comment_status")
